+++ @xen-orchestra/disk-cli/src/commands/info.mts
@@ -1,11 +1,11 @@
 import Disposable from 'promise-toolbox/Disposable'
 import { getSyncedHandler } from '@xen-orchestra/fs'
-import { openDisk } from '@xen-orchestra/backups/disks/openDisk.mjs'
+import { openDisposableDisk } from '@xen-orchestra/backups/disks'
 import { formatBytes } from '../utils.mjs'
 
 export async function infoCommand(handlerUrl: string, diskPath: string, _extraArgs: string[]): Promise<void> {
   await Disposable.use(getSyncedHandler({ url: handlerUrl }), async handler => {
-    await Disposable.use(openDisk(handler, diskPath), async disk => {
+    await Disposable.use(openDisposableDisk({ handler, path: diskPath }), async disk => {
       const isDifferencing = disk.isDifferencing()
       const virtualSize = disk.getVirtualSize()
       const blockSize = disk.getBlockSize()
